Add unit tests for dependency layers and initialized starter file parser tests
INCOMPLETE: Need to write up Starterfiles and hardcode their expected instances. This is a big integration test

This commit introduces the testing suite for critical functions in the 'startout' package. Two test files have been created for the purpose: `test_starterfile_utilities.py` and `test_starterfile_parser.py`. The former tests the creation of dependency layers with/without dependencies, missing dependencies, and circular dependencies. The latter is the start of a suite which validates the parsing of different starter files.